Moving jit type utils to legacy
