+++ cpp/src/binaryop/binaryop.cpp
@@ -30,7 +30,7 @@
 #include "compiled/binary_ops.hpp"
 
 #include <jit/launcher.h>
-#include <jit/type.h>
+#include <jit/legacy/type.h>
 #include <jit/parser.h>
 #include "jit/code/code.h"
 #include "jit/util/type.h"
